# Conflicts:
#	common/src/java/org/apache/hadoop/hive/conf/Constants.java
#	pom.xml
#	ql/src/java/org/apache/hadoop/hive/ql/optimizer/calcite/rules/views/HiveMaterializedViewRule.java
#	ql/src/java/org/apache/hadoop/hive/ql/parse/CalcitePlanner.java

+++ ql/src/java/org/apache/hadoop/hive/ql/optimizer/calcite/reloperators/HiveGroupingID.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -33,7 +33,7 @@ public class HiveGroupingID extends SqlAbstractGroupFunction {
   private HiveGroupingID() {
     super(VirtualColumn.GROUPINGID.getName(),
             SqlKind.OTHER,
-            ReturnTypes.INTEGER,
+            ReturnTypes.BIGINT,
             InferTypes.BOOLEAN,
             OperandTypes.NILADIC,
             SqlFunctionCategory.SYSTEM);
